added full testing for todos

+++ server/server.js
@@ -1,3 +1,6 @@
+require("../config/config");
+
+
 var express = require("express");
 var {mongoose} = require(".././db/mongoose");
 var {Todo} = require(".././db/models/Todo");
@@ -8,12 +11,11 @@ var _ = require("underscore");
 
 
 var app = express();
-var port = process.env.PORT || 8080;
 
 
-process.env.NODE_ENV = process.env.NODE_ENV || "development";
 
-console.log(process.env.NODE_ENV);
+
+
 
 app.use(bodyParser.json());
 
@@ -22,10 +24,23 @@ app.post("/todos", (req,res) => {
 
   var newTodo = new Todo(req.body);
 
+  if(req.body._id){
+    Todo.findById(new ObjectID(req.body._id)).then((doc) => {
+      if(doc){
+        return res.status(400).send("there is already a todo with this id");
+      }
+    }, (e) => {
+      return res.status(400).send(`someting went wrong : ${e}`);
+    });
+    if(!ObjectID.isValid(req.body._id)){
+      return res.status(400).send("id is not valid");
+    }
+  };
+
   newTodo.save().then(() => {
-    res.redirect("/todos");
+    return res.redirect("/todos");
   }).catch((e) => {
-    res.status(400).send(`someting went wrong : ${e}`);
+    return res.status(400).send(`someting went wrong : ${e}`);
   });
 
 });
@@ -34,7 +49,7 @@ app.post("/todos", (req,res) => {
 app.get("/todos", (req,res) => {
 
   Todo.find().then((todos) => {
-    res.status(200).send(todos);
+    res.status(200).send({todos});
   }).catch((e) => {
     console.log("something went wrong ",e);
     res.status(400).send("error while fetching data from db");
@@ -52,7 +67,6 @@ app.get("/todos/:id", (req,res) => {
 
   Todo.findById(new ObjectID(id)).then((doc) => {
 
-    console.log(doc);
     if(!doc){
       return res.status(404).send(`couldn't get todo with this id`);
     }
@@ -66,7 +80,8 @@ app.get("/todos/:id", (req,res) => {
 
 app.delete("/todos/:id",(req,res) => {
 
-
+  if(!ObjectID.isValid(req.params.id))
+    return res.status(404).send(`id has bad format`);
 
   Todo.findByIdAndRemove(new ObjectID(req.params.id)).then((doc) => {
       if(!doc)
@@ -79,7 +94,15 @@ app.delete("/todos/:id",(req,res) => {
 });
 
 app.patch("/todos/:id", (req,res) => {
+
+
+  if(!ObjectID.isValid(req.params.id))
+    return res.status(404).send("bad format of _id");
+
+
   var body = _.pick(req.body,["text","complited"]);
+
+
 
   if(_.isBoolean(body.complited) && body.complited) {
     body.complitedAt = Date.now();
@@ -107,8 +130,8 @@ app.get("*",(req,res) => {
   res.redirect("/todos");
 });
 
-app.listen(port, () => {
-  console.log("server listening at port ", port);
+app.listen(process.env.PORT, () => {
+  console.log("server listening at port ", process.env.PORT);
 });
 
 module.exports = {app};
